requirements.txt
Django==1.7.5
Pillow==2.7.0
djangorestframework==3.0.5
psycopg2==2.6
